Fix cli tests (move back to tests dir)
