+++ src/test_utils.rs
@@ -4,6 +4,8 @@ use http::StatusCode;
 use wiremock::matchers::path;
 use wiremock::{Mock, MockServer, ResponseTemplate};
 
+// TODO: used in cli tests (as duplicate)
+#[allow(unused)]
 pub(crate) async fn get_mock_server<S>(response_code: S) -> MockServer
 where
     S: Into<StatusCode>,
